Move SummaryMeasurementTable
Also extract inner classes to help maintainability
