+++ qupath-gui-fx/src/main/java/qupath/lib/gui/tools/MeasurementExporter.java
@@ -43,7 +43,7 @@ import org.slf4j.LoggerFactory;
 import com.google.common.collect.LinkedListMultimap;
 import com.google.common.collect.Multimap;
 
-import qupath.lib.gui.commands.SummaryMeasurementTable;
+import qupath.lib.gui.measure.ui.SummaryMeasurementTable;
 import qupath.lib.gui.commands.SummaryMeasurementTableCommand;
 import qupath.lib.gui.measure.ObservableMeasurementTableData;
 import qupath.lib.gui.prefs.PathPrefs;
